Extracting IntListVariantGenerator to its own file
