+++ gazeplay-games-commons/src/main/java/net/gazeplay/commons/gamevariants/generators/IGameVariantGenerator.java
@@ -1,0 +1,14 @@
+package net.gazeplay.commons.gamevariants.generators;
+
+import net.gazeplay.GameSpec;
+
+import java.util.Set;
+
+public interface IGameVariantGenerator {
+
+    default String getVariantChooseText() {
+        return "Choose Game Variant";
+    }
+
+    Set<GameSpec.GameVariant> getVariants();
+}
