+++ gazeplay-games/src/main/java/net/gazeplay/games/cups/CupsBallsGameVariantGenerator.java
@@ -1,8 +1,8 @@
 package net.gazeplay.games.cups;
 
-import net.gazeplay.GameSpec;
+import net.gazeplay.commons.gamevariants.generators.IntListVariantGenerator;
 
-public class CupsBallsGameVariantGenerator extends GameSpec.IntListVariantGenerator {
+public class CupsBallsGameVariantGenerator extends IntListVariantGenerator {
 
     public CupsBallsGameVariantGenerator() {
         super("Choose number of cups", 3, 5);
