+++ gazeplay-games/src/main/java/net/gazeplay/games/order/OrdersGameVariantGenerator.java
@@ -1,8 +1,8 @@
 package net.gazeplay.games.order;
 
-import net.gazeplay.GameSpec;
+import net.gazeplay.commons.gamevariants.generators.IntListVariantGenerator;
 
-public class OrdersGameVariantGenerator extends GameSpec.IntListVariantGenerator {
+public class OrdersGameVariantGenerator extends IntListVariantGenerator {
 
     public OrdersGameVariantGenerator() {
         super("Choose number of targets", 3, 5, 7);
